Create a blog post describing the creation of the i18n-babel-plugin
Summary:
This will be the second	blog post on our shiny new engineering blog.
It describes the creation of i18n-babel-plugin which replaces our
jsxtransform plugin.  The reason why we did this work is that we'd
like to switch to babel for compiling our JavaScript code b/c babel
supports more ES6/ES7.

Test Plan: load http://localhost:9103/ and	make sure the footnotes	work and there are no errors.

Reviewers: james, jordan

Differential Revision: https://phabricator.khanacademy.org/D18518