Fixed some layout problems with nav buttons.
Since the floated element wasn't affecting the height of its container,
we had some weird margins at the bottom of the page.

So if the right button was super tall and the left button was short or
non-existant, we wouldn't have any padding at the bottom of the page at
all!