Don't break when there's no upcoming post
Summary:
I don't think there's anyone actually after me in the publish queue,
so when I publish my post tomorrow I'll set `upcoming_post = None`.
That currently breaks the sidebar. This patch fixes that.

Test Plan:
At the bottom of `src/info.py`,
change the value of `upcoming_post` to `None`.
Build and make sure that the app doesn't break
and that no "upcoming post" is displayed in the sidebar.