Add a section on using i18nize-templates to extract out javascript.
Summary:
It makes the blog post even longer, but I think is worthwhile since it
shows that this tool is useful for things that are not explicitly
i18n-related.

I also added a 'summary' section at the end, so the blog post ends
less abruptly.  It still doesn't have a good kicker though. :-(

I wanted to add a non-image files to the blog post, so I created a new
supporting-files directory.  I had to modify the gulpfile to process
it.  The results *seem* right, but it will be hard to tell until it's
fully published.

Test Plan:
Ran
   make serve
in the engblog directory and looked at
   http://localhost:9103/
in my browser.  I made sure the page looked right, and also that
clicking on the 'here' link properly took me to
   http://localhost:9103/supporting-files/js_in_html.py
and that that page rendered properly as well.

Reviewers: johnsullivan

Subscribers: davidwang

Differential Revision: https://phabricator.khanacademy.org/D21691