Remove the lato link, just use a png for the instance hours graph
Summary:

The kotlin instance hours graph is still not showing up correctly if
you don't have lato installed. (Presumably I needed to put the <link>
in the <head>, which we can't configure on a per-post basis.)

This commit just switches us from an svg to a png to work around this
font issue. I also removed the font link tag that did not successfully
solve this.

Test Plan:
- `make serve`
- see that the image still shows up in the kotlin post
- verify I loaded it from png

Auditors: drosile, kevindangoor